refactor: consolidate duplicate isRecord, sleep, and truncateText
isRecord (16 copies → src/utils/type-guards.ts):
- 13 copies used the canonical form (excludes null and arrays)
- 3 copies were loose variants (allowed arrays) — all safe to upgrade
  since they were guarded by preceding Array.isArray checks or used
  in contexts where array behavior is equivalent
- channels/account-config.ts re-exports from the shared util

sleep (3 identical copies → src/utils/sleep.ts):
- bash-output.ts, task.ts, file-lock.ts all had character-for-character
  identical implementations

truncateText (4 copies → src/cli/helpers/truncate-text.ts):
- AgentInfoBar had the correct Unicode-aware version (stringWidth)
- McpSelector, SlashCommandAutocomplete, MessageSearch used text.length
  (incorrect for wide/emoji chars) — all upgraded to the Unicode-aware version
