+++ src/agent/modify.ts
@@ -11,6 +11,7 @@ import type { Conversation } from "@letta-ai/letta-client/resources/conversation
 import { getBackend } from "@/backend";
 import { OPENAI_CODEX_PROVIDER_NAME } from "@/providers/openai-codex-provider";
 import { debugLog } from "@/utils/debug";
+import { isRecord } from "@/utils/type-guards";
 import { getModelContextWindow } from "./available-models";
 
 type ModelSettings =
@@ -21,10 +22,6 @@ type ModelSettings =
 
 function supportsDistinctAnthropicXHighEffort(modelHandle: string): boolean {
   return modelHandle.includes("claude-opus-4-7");
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === "object" && value !== null && !Array.isArray(value);
 }
 
 /**
