+++ src/backend/local/local-backend.ts
@@ -26,6 +26,7 @@ import {
 } from "@/backend/dev/pi-stream-adapter";
 import type { ProviderTurnInput } from "@/backend/dev/provider-turn-executor";
 import { ProviderTurnExecutor } from "@/backend/dev/provider-turn-executor";
+import { isRecord } from "@/utils/type-guards";
 import {
   estimateLocalMessageTokens,
   isLocalSlidingWindowCompactionPlanningError,
@@ -143,10 +144,6 @@ interface ResolvedLocalCompactionSettings {
   prompt?: string | null;
   clipChars?: number | null;
   slidingWindowPercentage: number;
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === "object" && value !== null && !Array.isArray(value);
 }
 
 function compactionSettingsRecord(
