+++ src/channels/account-config.ts
@@ -87,9 +87,8 @@ const customChannelAccountConfigAdapter: ChannelAccountConfigAdapter<ChannelAcco
     },
   };
 
-export function isRecord(value: unknown): value is Record<string, unknown> {
-  return !!value && typeof value === "object" && !Array.isArray(value);
-}
+import { isRecord } from "@/utils/type-guards";
+export { isRecord };
 
 export function getChannelAccountConfigAdapter(
   channelId: string,
