+++ src/channels/plugin-registry.ts
@@ -1,6 +1,7 @@
 import { existsSync, readdirSync, readFileSync } from "node:fs";
 import { resolve, sep } from "node:path";
 import { pathToFileURL } from "node:url";
+import { isRecord } from "@/utils/type-guards";
 import { getChannelDir, getChannelsRoot } from "./config";
 import { CUSTOM_CHANNEL_CONFIG_SCHEMA } from "./custom/plugin";
 import type {
@@ -95,10 +96,6 @@ const FIRST_PARTY_CHANNEL_PLUGIN_REGISTRATIONS: Record<
 };
 
 const loadedUserPlugins = new Map<string, Promise<ChannelPlugin>>();
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return !!value && typeof value === "object" && !Array.isArray(value);
-}
 
 function isValidChannelId(value: string): boolean {
   return CHANNEL_ID_PATTERN.test(value);
