+++ src/utils/sleep.ts
@@ -1,0 +1,6 @@
+/**
+ * Returns a Promise that resolves after the given number of milliseconds.
+ */
+export function sleep(ms: number): Promise<void> {
+  return new Promise((resolve) => setTimeout(resolve, ms));
+}
